+++ lib/iqplayer.dart
@@ -1,4 +1,7 @@
 library iqplayer;
 
 export 'src/ui/iqscreen.dart';
-export 'package:video_player/video_player.dart';
+export 'src/utils/subtitle_provider.dart';
+export 'src/ui/iqparser.dart';
+export 'package:video_player/video_player.dart'
+    show VideoPlayerController, VideoPlayerValue;