+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0] - Spring.
 
-* TODO: Describe initial release.
+* Added `IQScreen` class that enable user to use a plyer as a screen.
+
+## [0.2.0] - Storm.
+
+* Added `SubtitleProvider` class that enable user to use a subtitle from files, assets, network, string.
+* Added `IQParser` class to display subtitle data.
+* Added `SubtitleBloc` class to use with `IQParser`.